backend/requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
SQLAlchemy==1.4.45
pydantic[email]~=1.10.7
envparse==0.2.0
psycopg2-binary==2.9.5
asyncpg==0.27.0
alembic==1.9.0
bcrypt==4.0.1
passlib==1.7.4
python-multipart==0.0.5
starlette~=0.22.0
python-jose==3.3.0
PyJWT==2.1.0